requirements.txt
# Core web framework
flask==3.1.0
python-dotenv==1.1.0
werkzeug==3.1.3
jinja2==3.1.6

# Document processing
pypdf==5.4.0
EbookLib==0.18
beautifulsoup4==4.13.4
lxml==5.3.2

# Vector database
pymilvus==2.5.6

# LangChain packages
langchain-core==0.3.54
langchain-community==0.3.21
langchain-text-splitters==0.3.8
langchain-openai==0.3.8  # Add this instead of using deprecated OpenAIEmbeddings

# LLM dependencies
openai==1.75.0
tiktoken==0.9.0

# Utilities
requests==2.32.3
numpy==2.0.2
pydantic==2.11.3
tenacity==9.1.2
